move helper functions to helper.js
